+++ src/main/kotlin/com/github/strindberg/emacsj/line/TransposeLinesHandler.kt
@@ -1,6 +1,11 @@
 package com.github.strindberg.emacsj.line
 
+import com.github.strindberg.emacsj.EmacsJCommandListener
+import com.github.strindberg.emacsj.mark.MarkHandler
+import com.github.strindberg.emacsj.universal.UniversalArgumentHandler
+import com.github.strindberg.emacsj.universal.universalCommandNames
 import com.github.strindberg.emacsj.word.substring
+import com.github.strindberg.emacsj.word.text
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Caret
 import com.intellij.openapi.editor.Editor
@@ -15,16 +20,31 @@ class TransposeLinesHandler : EditorWriteActionHandler() {
     override fun executeWriteAction(editor: Editor, editorCaret: Caret?, dataContext: DataContext) {
         val caret = editor.caretModel.primaryCaret
         val document = editor.document
-        val lineNum = document.getLineNumber(caret.offset)
+        val currentLineNumber = document.getLineNumber(caret.offset)
 
-        if (lineNum > 0) {
-            val previousLineStart = document.getLineStartOffset(lineNum - 1)
-            val currentLineStart = document.getLineStartOffset(lineNum)
-            val currentLineEnd = document.getLineEndOffset(lineNum)
+        val replaceLineNumber = if (EmacsJCommandListener.lastCommandName in universalCommandNames) {
+            if (UniversalArgumentHandler.lastArgument == 0) {
+                MarkHandler.peek(editor)?.caretPosition?.let { position -> document.getLineNumber(position) } ?: (currentLineNumber - 1)
+            } else {
+                currentLineNumber - UniversalArgumentHandler.lastArgument
+            }
+        } else {
+            currentLineNumber - 1
+        }
 
+        if (replaceLineNumber >= 0) {
+            val replaceLineStart = document.getLineStartOffset(replaceLineNumber)
+            val replaceLineEnd = document.getLineEndOffset(replaceLineNumber)
+            val currentLineStart = document.getLineStartOffset(currentLineNumber)
+            val currentLineEnd = document.getLineEndOffset(currentLineNumber)
+
+            val replaceLine = document.substring(replaceLineStart, replaceLineEnd)
             val currentLine = document.substring(currentLineStart, currentLineEnd)
-            document.deleteString(currentLineStart, currentLineEnd + 1)
-            document.insertString(previousLineStart, currentLine + "\n")
+
+            document.replaceString(currentLineStart, minOf(editor.text.length, currentLineEnd + 1), replaceLine + "\n")
+            document.replaceString(replaceLineStart, replaceLineEnd + 1, currentLine + "\n")
+
+            editor.caretModel.moveToOffset(currentLineEnd + 1)
         }
     }
 }